+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## Added
+
+- Support timeout on deletion.
+
 ## [4.10.1] - 2022-06-22
 
 ### Fixed